\name{nca.example}
\docType{data}
\alias{nca.example}
\title{NCA example data with 2 conditions and 1 output}
\description{
This data set has Individualism and Risk taking	as conditions,
and Innovation performance as the output for 28 countries.
}
\usage{data(nca.example)}
\format{A matrix containing 28 observations, incl. headers and row names.}
\keyword{datasets}