\name{nca.example2}
\docType{data}
\alias{nca.example2}
\title{NCA example data with 3 conditions and 1 outcome}
\description{
This data set has Contractual detail, Goodwill trust, and Competence trust
as conditions, and Innovation as outcome for 48 buyer-supplier relationships.
See Table 2 in: Van der Valk, W., Sumo, R., Dul, J., & Schroeder, R. G. (2016).
When are contracts and trust necessary for innovation in buyer-supplier relationships?
A necessary condition analysis.
Journal of Purchasing and Supply Management, 22(4), 266-277.
}
\usage{data(nca.example2)}
\format{A matrix containing 48 observations, incl. headers.}
\keyword{datasets}