% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca.R
\name{nca}
\alias{nca}
\title{neighborhood component analysis}
\usage{
nca(x, labels, A_init = diag(ncol(x)), N_iter = 100, learning_rate = 0.01)
}
\arguments{
\item{x}{numeric data that can be coerced into a numeric matrix}

\item{labels}{\code{factor} with the "correct" labels for each row in \code{x}}

\item{A_init}{start scale matrix \code{A}}

\item{N_iter}{\code{numeric} number of iterations}

\item{learning_rate}{rate in which change are made to matrix \code{A}}
}
\description{
Neighborhood Component Analysis
}

