\name{point.color}
\alias{point.color}
\docType{data}
\title{parameter defining the point color in the plots}
\description{This will be used for the \bold{col} parameters of the plots, default is blue.\cr
Set before calling nca_output\cr
> point.color <- red\cr\cr
See \href{http://www.statmethods.net/advgraphs/images/points.png}{
http://www.statmethods.net/advgraphs/images/points.png} for more symbols
}
\keyword{parameter}
