\name{nca.example}
\docType{data}
\alias{nca.example}
\title{NCA example data with 2 independent and 1 dependent variables}
\description{
This data set has Individualism and Risk taking	as independent variables,
and Innovation performance as the dependent variable for 28 countries.
}
\usage{data(nca.example)}
\format{A matrix containing 28 observations, incl. headers and row names.}
\keyword{datasets}