% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{nca_loss}
\alias{nca_loss}
\alias{loss_inaccuracy}
\alias{loss_sq_error}
\alias{loss_abs_error}
\alias{loss_tolerance}
\title{Loss functions}
\usage{
loss_inaccuracy(x, y, ...)

loss_sq_error(x, y, ...)

loss_abs_error(x, y, ...)

loss_tolerance(x, y, tol, ...)
}
\arguments{
\item{x, y}{Vectors whose loss to calculate}

\item{...}{Other arguments (not in use)}

\item{tol}{A tolerance below which things are considered neighbors and above
which is considered error}
}
\description{
Loss functions
}
