\name{point.type}
\alias{point.type}
\docType{data}
\title{parameter defining the plotting symbol in the plots}
\description{This will be used for the \bold{pch} parameter of the plots, default is 21.\cr
Set before calling nca_output\cr
  > point.type <- 22\cr\cr
See \href{http://www.statmethods.net/advgraphs/images/points.png}{
  http://www.statmethods.net/advgraphs/images/points.png} for more symbols
}
\keyword{parameter}
